Add auto dependency resolution, add generation of typing stubs
With these changes .NET Core SDK is required to be installed to build
the wheel.

Added some new scripts, which automate the package generation.
`init_itextpy_package.py` generates the itextpy package directory. This
is the package, which contains all the .NET binaries and the load
method for referencing them.

The dependency tree is resolved via a dummy .NET Core project, which has
the iText libraries as its dependencies. With how `dotnet publish`
works, it doesn't allow us to just get everything a user would need in
one `publish`, as, by default, it would just ignore platform-specific
binaries entirely.

This is where the script comes in. In the nutshell, it just calls `dotnet
publish` for a variety of common platforms, and then just collects and
merges the binaries via the script into a neat `binaries` tree, which now
has the additional directories for OS-specific binaries. At the same time,
the script generates the `load` function, which automatically adds all
the references based on the platform Python is running at.

As for stubs, they are auto-generated via
https://github.com/MHDante/pythonnet-stub-generator. Script for
generating them is `init_stubs_package.py`. It will generated another
package, `iText-stubs` via the tool after the base `itextpy` package has
been initialized. This package enables IDE support for `itextpy`.