Add binary patch to fix .NET Core issue in itext.io.dll
There is a bug in 9.1.0, causing an exception in
ResourceUtil::LoadITextResourceAssemblies, when using .NET Core under
Python.NET. The BaseDirectory is set to an empty string in such cases,
which iText doesn't handle properly right now.

For now adding a small binary patch, so that you can experiment easily
without recompiling iText with the fix yourself.