Make AbstractPdfDocumentEventHandler usable
Annoyingly, you cannot override protected methods in Python.NET. And
this is a big problem for AbstractPdfDocumentEventHandler, as you need
to override an `abstract protected` method.

Relevant issues:
- https://github.com/pythonnet/pythonnet/issues/2213
- https://github.com/pythonnet/pythonnet/issues/2571

I've tried making a Python.NET fork, which allows you to override:
- https://github.com/Eswcvlad/pythonnet/tree/protected-override

But since the issue has been hanging for a while, I doubt it will be
fixed soon. And making a dependency on a fork is not great...

Changing this on iText side is not good as well, as it makes the C# API
worse...

For now, I've patched all AbstractPdfDocumentEventHandler
OnAcceptedEvent methods in iText to be public, but this is a very bad
solution, as:
- You would need to track any new subclasses on updates and patch them
  as well.
- You would need to update these patches for every release.
- This only handles one method, while this is a global problem.
- If you add another library, which also derives from
  AbstractPdfDocumentEventHandler, then it will not work, if
  OnAcceptedEvent remained "protected" there.

But it is better than nothing...