Depend on better .NET type stubs
I've published auto-generated type stubs for .NET Standard stubs, which
work better than the dotnet-stubs one. So we can rely on that instead.

Practically speaking, there is no need to depend on stubs here, as only
the developer needs them, not the final product. But this is convenient
to have in an example repo.