Pin .NET transitive dependencies versions
Somewhat surprising, there is still no way in a modern .NET project to
get the highest compatible version of packages instead of the lowest
ones... For now using transitive package pinning to get newer versions.
Hopefully by the time we will need to update this, it won't be required
anymore.