Add a fallback for DLL hell situations
Here is an example, when DLL hell could appear. If you, for example,
have .NET 8 and try to load System.Text.Encodings.Web version 9.0.5,
it will not load. This happens because .NET 8 has a bundled 8.0.0
version of System.Text.Encodings.Web, which get loaded automatically
and conflicts, when we try to use the later one.

Now, as a fallback, if we detect a FileLoadException, instead we will
just try to load the library from the default location and output a
warning for the user.